requirements.txt
absl-py==1.3.0
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
astunparse==1.6.3
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
cached-property==1.5.2
cachetools==5.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
comm==0.1.2
confluent-kafka==2.0.2
cycler==0.11.0
debugpy==1.6.4
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.16.2
flatbuffers==22.12.6
fonttools==4.38.0
fqdn==1.5.1
future==0.18.3
gast==0.4.0
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
idna==3.4
oauthlib==3.2.2
py4j==0.10.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyspark==2.3.1
python-dateutil==2.8.2
python-twitter==3.5
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
traitlets==5.9.0
tweepy==4.12.1
typing-extensions==4.5.0
urllib3==1.26.14
webencodings==0.5.1
